requirements.txt
autopep8==2.0.1
aws-cdk-lib==2.66.0
awscli==1.27.78
boto3==1.26.78
constructs>=10.0.0,<11.0.0
flake8==6.0.0
isort==5.12.0
pep8==1.7.1
pip==23.0.1
yapf==0.32.0
